An alternative Block representation that gives O(1) access to the first and last node, which makes it easier to perform operations on these nodes.
With the new representation we can have

  blockSplit :: Block n C C -> (n C O, Block n O O, n O C)
  blockSplit (BlockCC f b t) = (f, b, t)

  blockJoin :: n C O -> Block n O O -> n O C -> Block n C C
  blockJoin f b t = BlockCC f b t

whereas in the old representation these would be O(n), and would need
an extra Maybe (because Blocks can't be empty).  This came up because
GHC needs these kinds of operations - search for uses of
blockToNodeList in compiler/cmm for some examples.  Many of these are
just picking apart blocks to get at the first or last nodes.

The simonmar-alternative-block-rep branch makes the representation
change to the extent required to compile GHC master and do performance
measurements.  Some functions that GHC does not require are commented
out.